Renames buynch of files
