+++ bvareader/helpers.py
@@ -1,0 +1,3 @@
+def flatten_list(list_of_lists):
+	return([item for sublist in list_of_lists for item in sublist])
+
